refactor: App.ts 파일 - dropdown event 핸들러 분리
